+++ src/axom/core/CMakeLists.txt
@@ -34,6 +34,7 @@ set(core_headers
     utilities/StringUtilities.hpp
     utilities/System.hpp
     utilities/Timer.hpp
+    utilities/Units.hpp
     utilities/Utilities.hpp
 
     ## numerics
@@ -108,6 +109,7 @@ set(core_sources
     utilities/FileUtilities.cpp
     utilities/StringUtilities.cpp
     utilities/System.cpp
+    utilities/Units.cpp
     utilities/Utilities.cpp
     ${PROJECT_BINARY_DIR}/axom/core/utilities/About.cpp
