+++ src/axom/core/utilities/Units.hpp
@@ -3,18 +3,14 @@
 // files for dates and other details.
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
-#ifndef AXOM_KLEE_UNITS_HPP
-#define AXOM_KLEE_UNITS_HPP
+#ifndef AXOM_CORE_UTILITIES_UNITS_HPP
+#define AXOM_CORE_UTILITIES_UNITS_HPP
 
 #include <string>
 
 namespace axom
 {
-namespace inlet
-{
-class Proxy;
-}
-namespace klee
+namespace utilities
 {
 /**
  * Units of length in which users can express lengths and in which client
@@ -41,22 +37,10 @@ enum class LengthUnit
  * Convert a string to a LengthUnit.
  *
  * \param unitsAsString the units as a string
- * \param path the Path where the length units were found in the document.
- * Used for error reporting.
  * \return the parsed units
- * \throws KleeError if the string does not represent known
- * units
+ * \throws std::invalid_argument if the string does not represent known units
  */
-LengthUnit parseLengthUnits(const std::string &unitsAsString, const std::string &path);
-
-/**
- * Convert a proxy to a LengthUnit.
- *
- * \param unitsAsProxy the units as a proxy
- * \return the parsed units
- * \throws KleeError if the string does not represent known units
- */
-LengthUnit parseLengthUnits(const inlet::Proxy &unitsAsProxy);
+LengthUnit parseLengthUnits(const std::string &unitsAsString);
 
 /**
  * Get the conversion factor to convert from the given source units to the target units.
@@ -97,6 +81,6 @@ void convertAll(T &values, LengthUnit sourceUnits, LengthUnit targetUnits)
   }
 }
 
-}  // namespace klee
+}  // namespace utilities
 }  // namespace axom
-#endif  // AXOM_KLEE_UNITS_HPP
+#endif  // AXOM_CORE_UTILITIES_UNITS_HPP
